tests/baseline/protocol-1.2/inspector/tracing/timeline-js/timeline-script-tag-2-expected.txt

CONSOLE MESSAGE: line 1: SCRIPT TAG
Tests the Timeline API instrumentation of a script tag with an external script.

EvaluateScript Properties:
{
    data : {
        columnNumber : <number>
        frame : <string>
        lineNumber : <number>
        url : .../inspector/tracing/timeline-js/timeline-script-tag-2.js
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "EvaluateScript"
}
